V - 3.2
Electricity and Magnetism tab:
Magnetism tab still needs to be fixed so that i can only enter either an R value (Radius) or an N value (Number of turns) to determine if we must use the straight wire formula or the solenoid formula for magnetic field strength as a result of Amperes law. This should allow us to determine what  graph to use (solenoid or straight line).  Work on it tmrw!!!!! Refer to tab 6 DS.