V - 4.5.4
It WAS KINDA NOT REALLY WOKRING (i got an input but it was completely wrong lmao). I tried to fix it, became worse cause now "error': 'Processing failed: not enough values to unpack " so the parsing is at an all time low, but ig atleast its still trying to parse? Finish tmrw w/ refrence to DS.